pyproject.toml
[tool.poetry]
name = "applimate"
version = "0.1.0"
description = "Applimate Backend API V1"
authors = ["AbdulAfeez <abdulafeez.adeyemo@kibo.school>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
django-rest-framework = "^0.1.0"
djangorestframework-simplejwt = "^5.3.1"
pyotp = "^2.9.0"
drf-social-oauth2 = "^2.1.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
